bot/rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "google-generativeai>=0.8.6",
    "langchain-text-splitters>=1.1.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "tabulate>=0.10.0",
]
